+++ test_projects/adventure/main.rb
@@ -3,6 +3,7 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __dir__)
 require 'rgame/game'
 
+require_relative 'debug_toggle'
 require_relative 'hero'
 require_relative 'room'
 require_relative 'shell'
@@ -11,6 +12,7 @@ WIDTH  = 640
 HEIGHT = 480
 
 ASSETS = File.expand_path('../../examples/assets', __dir__)
+Controls = RGame::Util::Controls
 
 game = RGame::Game.new(
   root: Shell.new,
@@ -18,7 +20,10 @@ game = RGame::Game.new(
   width: WIDTH,
   height: HEIGHT,
   media_root: ASSETS,
-  players: 2
+  players: 2,
+  input_map: RGame::Engine::InputMap.default.merge(
+    debug: { buttons: [Controls::KEY_F4, Controls::PAD_BACK] }
+  )
 )
 
 game.start